Added free scoring list
Also made agents spam the output with info about how long it thinks each
chain is. Current problems:
- 3 clusters are not counted as short chains.
- Forks with 2 short branches are considered a short chain once a branch
is sacrificed (this won't be a problem when evaluating, just when
playing it out for control).
- Problems when considering double boxing when opponent sacrifices in
the middle of a long chain or in a loop (not actually at double boxing
stage yet but need to note this)